+++ bdr_relcache.c
@@ -28,12 +28,6 @@
 #include "utils/catcache.h"
 #include "utils/fmgroids.h"
 #include "utils/inval.h"
-#if PG_VERSION_NUM >= 130000
-#include "common/jsonapi.h"
-#else
-#include "utils/jsonapi.h"
-#endif
-#include "utils/json.h"
 #include "utils/jsonb.h"
 #include "utils/rel.h"
