+++ compat/15/bdr_compat.h
@@ -1,29 +1,16 @@
 #ifndef PG_BDR_COMPAT_H
 #define PG_BDR_COMPAT_H
 
-#include "access/xlogrecovery.h"
 #include "access/heapam.h"
 #include "access/genam.h"
 #include "common/hashfn.h"
 #include "access/htup_details.h"
+#include "utils/syscache.h"
 
-static inline TimeLineID
-GetTimeLineID (void)
-{
-	TimeLineID current_timeline;
-
-	if (RecoveryInProgress())
-		(void) GetXLogReplayRecPtr(&current_timeline);
-		else
-			current_timeline = GetWALInsertionTimeLine();
-
-	return current_timeline;
-}
-
-#define BDR_LOCALID_FORMAT_ARGS \
-	GetSystemIdentifier(), GetTimeLineID(), MyDatabaseId, EMPTY_REPLICATION_NAME
-
-//XXX should that be table_slot_create for >= 12 instead?
+/*
+ * XXX Should it be table_slot_create for >= 12 instead of
+ * ExecInitExtraTupleSlot?
+ */
 #define ExecInitExtraTupleSlotBdr(estate, a) \
 	ExecInitExtraTupleSlot(estate, a, &TTSOpsHeapTuple)
 
@@ -45,14 +32,27 @@ GetTimeLineID (void)
 
 #define TTS_TUP(slot) (((HeapTupleTableSlot *)slot)->tuple)
 
-#define BdrGetSysCacheOid GetSysCacheOid
-
 #define BdrGetSysCacheOid1 GetSysCacheOid1
 
 #define BdrGetSysCacheOid2 GetSysCacheOid2
 
-#define GetSysCacheOidError2(cacheId, oidcol, key1, key2) \
-	GetSysCacheOidError(cacheId, oidcol, key1, key2, 0, 0)
+/* GetSysCacheOid2 equivalent that errors out if nothing is found */
+static inline  Oid
+GetSysCacheOid2Error(int cacheId, AttrNumber oidcol,
+					 Datum key1, Datum key2)
+{
+	Oid			result;
+
+	result = GetSysCacheOid2(cacheId, oidcol, key1, key2);
+
+	if (result == InvalidOid)
+		elog(ERROR, "cache lookup failure in cache %d", cacheId);
+
+	return result;
+}
+
+#define BdrGetSysCacheOid2Error(cacheId, oidcol, key1, key2) \
+	GetSysCacheOid2Error(cacheId, oidcol, key1, key2)
 
 /* 2a10fdc4307a667883f7a3369cb93a721ade9680 */
 #define getObjectDescription(object) getObjectDescription(object, false)
